bot/requirements.txt
discord.py[voice] @ git+https://github.com/Rapptz/discord.py
python-dotenv>=1.0.0
webrtcvad>=2.0.10
aiohttp>=3.9.0
PyNaCl>=1.5.0
pydub>=0.25.1
